Implementando rolagem de lista e refatoção do projeto em pacotes
