+++ app/src/main/java/com/marcelo/loteriadossonhos/model/MainItem.kt
@@ -1,0 +1,11 @@
+package com.marcelo.loteriadossonhos.model
+
+import androidx.annotation.DrawableRes
+import androidx.compose.ui.graphics.Color
+
+data class MainItem(
+    val id: Int,
+    val name: String,
+    val color: Color,
+    @DrawableRes val icon: Int
+)
